text_adventure/app/requirements.txt
langchain-community==0.0.19
langchain-core==0.1.21
langchain-openai==0.0.5
langchain==0.1.5
langsmith==0.0.87
python-dotenv==1.0.1
openai==1.11.1
streamlit==1.29.0
cassandra-driver==3.29.1
cassio==0.1.5
astrapy==1.0.0
astroid==2.15.5
asttokens==2.2.1
astunparse==1.6.3
requests==2.31.0
requests-oauthlib==1.3.1
pypdf
tiktoken==0.6.0
click==8.1.7
numpy==1.23.5
dataclasses-json==0.6.4
fastjsonschema==2.17.1
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
orjson==3.10.3
python-json-logger==2.0.7
